Don't serve movies with their original file names
This uses the newly assigned UUID on movies as a "secret" to access a
movie file. This avoid leaking movies to unauthenticated people since
they could try guessing filenames (see #5).

To avoid serving movies through django, we set the `X-Accel-Redirect`
header to what nginx calls an internal route. That means that the route
cannot only be served to nginx itself. More info on that here:
https://www.nginx.com/resources/wiki/start/topics/examples/x-accel/

When presented with that header, nginx will ignore the rest of the
response and instead serve what's behind the specified URI. Doing it
this way allows us to keep supporting sending movies through django
(very useful for dev purposes) and keep supporting instances that want
those URLs to not be masked.

Using django's `serve` function sets all the proper headers
so the downloaded file gets its original name.

Note that the route to get a movie is still unauthenticated because
that's useful if you want to use an external viewer like mpv or VLC.

Fixes #5